Move `MatrixCredentials` to `auth/types` together to `MatrixClientCredentials`
