+++ src/auth/oidc/types.ts
@@ -19,28 +19,6 @@
 import Joi from 'joi';
 import { IdTokenClaims } from 'oidc-client-ts';
 
-export type MatrixClientCredentials = {
-  // url of the homeserver selected during login
-  homeserverUrl: string;
-  // identity server url as discovered during login
-  identityServerUrl?: string;
-  // accessToken gained from OIDC token issuer
-  accessToken: string;
-  // refreshToken gained from OIDC token issuer, when falsy token cannot be refreshed
-  refreshToken?: string;
-};
-
-export const matrixClientCredentialsSchema =
-  Joi.object<MatrixClientCredentials>({
-    homeserverUrl: Joi.string().uri().required(),
-    identityServerUrl: Joi.string().uri(),
-    accessToken: Joi.string(),
-    refreshToken: Joi.string(),
-  });
-
-/**
- * Borrowed from {@link https://github.com/matrix-org/matrix-react-sdk/blob/79c50db00993a97a0b6b8c3df02b8eec4e6cb21a/src/utils/oidc/authorize.ts#L80}
- */
 export type OidcCredentials = {
   // this client's id as registered with the OIDC issuer
   clientId: string;
