+++ src/state/index.ts
@@ -18,7 +18,6 @@
 
 export * from './Application';
 export { Credentials } from './Credentials';
-export type { MatrixCredentials } from './Credentials';
 export type * from './types';
 export { useDistinctObserveBehaviorSubject } from './useDistinctObserveBehaviorSubject';
 export { LoggedInProvider, useLoggedIn } from './useLoggedIn';
